feat : [ Product ] 도메인 구현

+++ src/main/java/kr/hhplus/ecommerce/application/product/dto/ProductResult.java
@@ -43,7 +43,7 @@ public record ProductResult() {
             Long optionId,
             String optionValue,
             Long price,
-            Integer stock
+            Long stock
     ) {
         public static Option from(Option option) {
             return Option.builder()
